+++ Magic/CHANGELOG.md
@@ -32,6 +32,7 @@
  - Add destructible_materials and indestructible_materials options to ChangeContext action
  - Add `until` parameter to Skip action, to skip sub-actions until some amount of time has passed
  - Add attributes: target_location_x, target_location_y, target_location_z
+ - Add /mauto enable/disable commands to quickly turn on/off automata
 
 # 8.3.3